Move preview query into new `BeeperStore` class
