+++ synapse/handlers/sync.py
@@ -40,7 +40,6 @@ from synapse.handlers.relations import BundledAggregations
 from synapse.logging.context import current_context
 from synapse.logging.opentracing import SynapseTags, log_kv, set_tag, start_active_span
 from synapse.push.clientformat import format_push_rules_for_user
-from synapse.storage.database import LoggingTransaction
 from synapse.storage.databases.main.event_push_actions import RoomNotifCounts
 from synapse.storage.databases.main.roommember import extract_heroes_from_room_summary
 from synapse.storage.roommember import MemberSummary
@@ -57,7 +56,6 @@ from synapse.types import (
     UserID,
 )
 from synapse.util.async_helpers import concurrently_execute
-from synapse.util.caches.descriptors import cached
 from synapse.util.caches.expiringcache import ExpiringCache
 from synapse.util.caches.lrucache import LruCache
 from synapse.util.caches.response_cache import ResponseCache, ResponseCacheContext
@@ -1295,50 +1293,6 @@ class SyncHandler:
                 sync_config.user.to_string(),
             )
 
-    @cached(max_entries=50000, iterable=True)
-    async def beeper_preview_for_room_id_and_user_id(
-        self, room_id: str, user_id: str, to_key: RoomStreamToken
-    ) -> JsonDict:
-        res = {}
-
-        def _beeper_preview_for_room_id_and_user_id(txn: LoggingTransaction) -> None:
-            sql = """
-            SELECT e.event_id, e.origin_server_ts
-                FROM events AS e \
-                LEFT JOIN redactions as r \
-                ON e.event_id = r.redacts
-            WHERE e.stream_ordering <= ? AND e.room_id = ? \
-            AND (e.type = "m.room.message" \
-            OR e.type = "m.room.encrypted" \
-            OR e.type = "m.reaction") \
-            AND r.redacts IS NULL \
-            AND CASE WHEN (e.type = "m.reaction") \
-                THEN (SELECT ? = ee.sender AND ee.event_id NOT IN (SELECT redacts FROM redactions WHERE redacts = ee.event_id) FROM events as ee \
-                    WHERE ee.event_id = (SELECT er.relates_to_id FROM event_relations AS er \
-                    WHERE er.event_id = e.event_id)) ELSE (true) END
-            ORDER BY stream_ordering DESC
-            LIMIT 1
-            """
-
-            txn.execute(
-                sql,
-                (
-                    to_key.stream,
-                    room_id,
-                    user_id,
-                ),
-            )
-            for event_id, origin_server_ts in txn:
-                res["event_id"] = event_id
-                res["origin_server_ts"] = origin_server_ts
-
-        await self.store.db_pool.runInteraction(
-            "beeper_preview_for_room_id_and_user_id",
-            _beeper_preview_for_room_id_and_user_id,
-        )
-
-        return res
-
     async def generate_sync_result(
         self,
         sync_config: SyncConfig,
@@ -2487,7 +2441,7 @@ class SyncHandler:
 
                 if self.hs_config.experimental.server_side_room_preview_enabled:
                     preview: JsonDict = (
-                        await self.beeper_preview_for_room_id_and_user_id(
+                        await self.store.beeper_preview_for_room_id_and_user_id(
                             room_id=room_id, user_id=user_id, to_key=now_token.room_key
                         )
                     )
