+++ synapse/storage/databases/main/__init__.py
@@ -30,6 +30,7 @@ from synapse.types import JsonDict, get_domain_from_id
 
 from .account_data import AccountDataStore
 from .appservice import ApplicationServiceStore, ApplicationServiceTransactionStore
+from .beeper import BeeperStore
 from .cache import CacheInvalidationWorkerStore
 from .censor_events import CensorEventsStore
 from .client_ips import ClientIpWorkerStore
@@ -125,6 +126,7 @@ class DataStore(
     CacheInvalidationWorkerStore,
     LockStore,
     SessionStore,
+    BeeperStore,
 ):
     def __init__(
         self,
